+++ projects/Mallard/src/components/front/items/item.tsx
@@ -31,7 +31,7 @@ import {
     useCardBackgroundStyle,
     ItemSizes,
     PageLayoutSizes,
-} from '../helpers'
+} from '../helpers/helpers'
 import {
     setScreenPositionOfItem,
     getScreenPositionOfItem,
