+++ projects/Mallard/src/components/layout/ui/with-breakpoints.tsx
@@ -1,14 +1,11 @@
 import React, { useState, ReactElement } from 'react'
 import { View, LayoutRectangle, StyleSheet } from 'react-native'
-import { getClosestBreakpoint } from 'src/theme/breakpoints'
+import { getClosestBreakpoint, BreakpointList } from 'src/theme/breakpoints'
 
 const WithBreakpoints = ({
     children,
 }: {
-    children: {
-        0: (l: LayoutRectangle) => ReactElement
-        [fromSize: number]: (l: LayoutRectangle) => ReactElement
-    }
+    children: BreakpointList<(l: LayoutRectangle) => ReactElement>
 }) => {
     const [maxSize, setMaxSize] = useState<keyof typeof children>(0)
     const [metrics, setMetrics] = useState<LayoutRectangle | null>(null)
