+++ projects/Mallard/src/theme/breakpoints.ts
@@ -3,6 +3,11 @@ export enum Breakpoints {
     tabletLandscape = 1000,
 }
 
+export interface BreakpointList<T> {
+    0: T
+    [fromSize: number]: T
+}
+
 export const getClosestBreakpoint = (breakpoints: number[], size: number) => {
     let max = 0
     for (let key of breakpoints) {
@@ -12,3 +17,14 @@ export const getClosestBreakpoint = (breakpoints: number[], size: number) => {
     }
     return max
 }
+
+export const pickClosestBreakpoint = <T>(
+    breakpoints: BreakpointList<T>,
+    size: number,
+) =>
+    breakpoints[
+        getClosestBreakpoint(
+            (Object.keys(breakpoints) as unknown[]) as number[],
+            size,
+        )
+    ]
